[FLINK-13632] Port EnumSerializer upgrade test to TypeSerializerUpgradeTestBase
